+++ CHANGELOG.md
@@ -7,3 +7,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.1.0 - 2023-12-01]
 ### Added
 - First version of the project
+- Spring Application